+++ tests/conftest.py
@@ -1,18 +1,11 @@
-"""PyTest configuration.
-
-Note: the calcipy imports are required for a nicer test HTML report
-
-"""
+"""PyTest configuration."""
 
 from pathlib import Path
-from typing import List, Tuple
 
 import pytest
-from calcipy.dev.conftest import pytest_configure  # noqa: F401
-from calcipy.dev.conftest import pytest_html_results_table_header  # noqa: F401
-from calcipy.dev.conftest import pytest_html_results_table_row  # noqa: F401
-from calcipy.dev.conftest import pytest_runtest_makereport  # noqa: F401
-from calcipy.dot_dict import ddict  # PLANNED: DDICT_TYPE
+from beartype import beartype
+from beartype.typing import List
+from calcipy.dot_dict import ddict
 from commitizen import defaults, git
 from commitizen.config import BaseConfig
 
@@ -21,6 +14,7 @@ from .constants import ANSWERS, COMMITS_DATA, TAGS
 
 
 @pytest.fixture()
+@beartype
 def fix_test_cache() -> Path:
     """Fixture to clear and return the test cache directory for use.
 
@@ -33,6 +27,7 @@ def fix_test_cache() -> Path:
 
 
 @pytest.fixture()
+@beartype
 def config() -> BaseConfig:
     """Create default configuration.
 
@@ -62,7 +57,7 @@ def config() -> BaseConfig:
         ),
     ],
 )
-def messages(request: List[Tuple[dict, str]]):  # -> DDICT_TYPE:
+def messages(request):  # noqa: RBT001
     """Fixture for raw answer and expected formatted commit message.
 
     Args:
@@ -76,6 +71,7 @@ def messages(request: List[Tuple[dict, str]]):  # -> DDICT_TYPE:
 
 
 @pytest.fixture()
+@beartype
 def gitcommits() -> List[git.GitCommit]:
     """Generate list of Git Commits.
 
@@ -94,6 +90,7 @@ def gitcommits() -> List[git.GitCommit]:
 
 
 @pytest.fixture()
+@beartype
 def tags() -> List[git.GitTag]:
     """Generate list of Git tTags.